Add SecureLogger with AES encrypted logging
