+++ AGENTS.md
@@ -185,6 +185,7 @@
  - [x] 8.2. Expand meta-learning (analysis of training outcomes, adjustment of packages, etc.)
  - [x] 8.3. Support version preview and rollback features
  - [x] 8.4. Document ability to scaffold new software projects
+ - [x] 8.5. Introduce SecureLogger with AES-encrypted logs
 
 ---
