+++ README.md
@@ -31,7 +31,8 @@ Initial structure for the autonomous polyglot code engineering system.
 - `knowledge_base/packages/` – curated guides loaded by the learning engine at startup.
 - `knowledge_base/offline_training/` – sample datasets used to seed offline models.
 - `data/` – runtime data generated by AI providers and the application. Older versions are archived under `data/versions/`.
-- `logs/` – log files for debugging and audit.
+- `logs/` – log files for debugging and audit. When `LOG_ENCRYPTION_KEY` is set
+  the contents are AES encrypted.
 - `code/` – workspace for generated or user‑provided code.
 - `src/` – .NET source projects:
   - `ASL.CodeEngineering.App` – WPF application hosting the UI.
@@ -52,10 +53,13 @@ data and logs:
   `data/` beside the executable).
 - `KB_DIR` – directory for generated summaries and other knowledge base
   files (defaults to `knowledge_base/`).
- - `LOGS_DIR` – directory for diagnostic logs (defaults to `logs/`). When this
-   directory cannot be written to, logs fall back to a `logs/` folder located
-   beside the executable.
- - `DISABLE_NETWORK_PROVIDERS` – set to `1` or `true` to remove providers that
+- `LOGS_DIR` – directory for diagnostic logs (defaults to `logs/`). When this
+  directory cannot be written to, logs fall back to a `logs/` folder located
+  beside the executable.
+- `LOG_ENCRYPTION_KEY` – optional key enabling AES-encrypted log files. When set,
+  `SecureLogger` stores each log with a prepended IV so the audit trail remains
+  confidential.
+- `DISABLE_NETWORK_PROVIDERS` – set to `1` or `true` to remove providers that
   require internet access from the dropdown. Use this to enforce a fully offline
   workflow where all prompts stay inside the project directory.
