+++ src/ASL.CodeEngineering.AI/AIProviderLoader.cs
@@ -95,11 +95,7 @@ public static class AIProviderLoader
         Debug.WriteLine($"{context}: {ex}");
         try
         {
-            var logsDir = Environment.GetEnvironmentVariable("LOGS_DIR") ??
-                          Path.Combine(AppContext.BaseDirectory, "logs");
-            Directory.CreateDirectory(logsDir);
-            var file = Path.Combine(logsDir, $"{context}_{DateTime.UtcNow:yyyyMMdd_HHmmss}.log");
-            File.WriteAllText(file, ex.ToString());
+            SecureLogger.Write(context, ex.ToString());
         }
         catch
         {
