+++ src/ASL.CodeEngineering.AI/ProcessRunner.cs
@@ -33,29 +33,6 @@ public static class ProcessRunner
 
     private static void Log(string prefix, string content)
     {
-        var logsDir = Environment.GetEnvironmentVariable("LOGS_DIR") ??
-                      Path.Combine(AppContext.BaseDirectory, "logs");
-
-        if (!TryWrite(logsDir))
-        {
-            var fallback = Path.Combine(AppContext.BaseDirectory, "logs");
-            if (fallback != logsDir)
-                TryWrite(fallback);
-        }
-
-        bool TryWrite(string dir)
-        {
-            try
-            {
-                Directory.CreateDirectory(dir);
-                var file = Path.Combine(dir, $"{prefix}_{DateTime.UtcNow:yyyyMMdd_HHmmss}.log");
-                File.WriteAllText(file, content);
-                return true;
-            }
-            catch
-            {
-                return false;
-            }
-        }
+        SecureLogger.Write(prefix, content);
     }
 }
